pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aleswitcher"
version = "2.0.1"
description = "AleSwitcher - Claude Code Account Switcher Desktop App"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "psutil>=5.9.0",
    "filelock>=3.12.0",
    "pywebview>=5.0",
    "pystray>=0.19",
    "Pillow>=10.0",
    "plyer>=2.1",
]

[project.gui-scripts]
ale_switcher = "ale_switcher.__main__:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools]
include-package-data = true
